extracted EventCollectorCoordinator and builder
